Happy path scenario test
