+++ package.json
@@ -51,6 +51,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react-calendar": "^3.5.0"
+    "@types/react-calendar": "^3.5.0",
+    "msw": "^0.39.2"
   }
 }
