+++ src/tests/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {act, fireEvent, render, screen} from '@testing-library/react';
-import LoginPage from "./pages/LoginPage";
+import LoginPage from "../pages/LoginPage";
 
 test('renders login button and form', () => {
   render(<LoginPage />);
